third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/import-attributes/unsupported-attribute-expected.txt

This is a testharness.js-based test.
FAIL Test that invalid module attribute leads to SyntaxError on window. assert_equals: expected 1 but got 2
Harness: the test ran to completion.
